CDEC-438: Move Shutdown from the sin bin to infra
