+++ infra/src/Pos/Infra/Shutdown.hs
@@ -1,5 +1,9 @@
 module Pos.Infra.Shutdown
-       ( module X
+       ( module Pos.Infra.Shutdown.Class
+       , module Pos.Infra.Shutdown.Logic
+       , module Pos.Infra.Shutdown.Types
        ) where
 
-import           Pos.Sinbin.Shutdown as X
+import           Pos.Infra.Shutdown.Class
+import           Pos.Infra.Shutdown.Logic
+import           Pos.Infra.Shutdown.Types
