+++ infra/src/Pos/Infra/Shutdown/Class.hs
@@ -1,5 +1,16 @@
+{-# LANGUAGE TypeFamilies #-}
+
 module Pos.Infra.Shutdown.Class
-       ( module X
+       ( HasShutdownContext(..)
        ) where
 
-import           Pos.Sinbin.Shutdown.Class as X
+import           Universum
+
+import           Control.Lens.Lens (lens)
+import           Pos.Infra.Shutdown.Types (ShutdownContext)
+
+class HasShutdownContext ctx where
+    shutdownContext :: Lens' ctx ShutdownContext
+
+instance HasShutdownContext ShutdownContext where
+    shutdownContext = lens identity (\_ x -> x)
