+++ infra/src/Pos/Infra/Shutdown/Logic.hs
@@ -1,5 +1,24 @@
 module Pos.Infra.Shutdown.Logic
-       ( module X
+       ( triggerShutdown
+       , waitForShutdown
        ) where
 
-import           Pos.Sinbin.Shutdown.Logic as X
+import           Universum
+
+import           Control.Concurrent.STM (check, readTVar, writeTVar)
+import           System.Wlog (WithLogger, logInfo)
+
+import           Pos.Infra.Shutdown.Class (HasShutdownContext (..))
+import           Pos.Infra.Shutdown.Types (ShutdownContext (..),
+                     shdnIsTriggered)
+
+triggerShutdown
+    :: (MonadIO m, MonadReader ctx m, WithLogger m, HasShutdownContext ctx)
+    => m ()
+triggerShutdown = do
+    logInfo "NODE SHUTDOWN TRIGGERED, WAITING FOR WORKERS TO TERMINATE"
+    view (shutdownContext . shdnIsTriggered) >>= atomically . flip writeTVar True
+
+-- | Wait for the shutdown var to be true.
+waitForShutdown :: ShutdownContext -> IO ()
+waitForShutdown (ShutdownContext v) = atomically (readTVar v >>= check)
