+++ infra/src/Pos/Infra/Shutdown/Types.hs
@@ -1,5 +1,17 @@
+{-# LANGUAGE TemplateHaskell #-}
+
 module Pos.Infra.Shutdown.Types
-       ( module X
+       ( ShutdownContext (..)
+       , shdnIsTriggered
        ) where
 
-import           Pos.Sinbin.Shutdown.Types as X
+import           Universum
+
+import           Control.Lens (makeLenses)
+
+data ShutdownContext = ShutdownContext
+    { _shdnIsTriggered :: !(TVar Bool)
+    -- ^ If this flag is `True`, then workers should stop.
+    }
+
+makeLenses ''ShutdownContext
