+++ sinbin/cardano-sl-sinbin.cabal
@@ -25,11 +25,6 @@ library
                         Pos.Sinbin.Reporting.Metrics
                         Pos.Sinbin.Reporting.Wlog
 
-                        Pos.Sinbin.Shutdown
-                        Pos.Sinbin.Shutdown.Class
-                        Pos.Sinbin.Shutdown.Logic
-                        Pos.Sinbin.Shutdown.Types
-
                         Pos.Sinbin.StateLock
 
                         Pos.Sinbin.Util.JsonLog.Events
