--- sinbin/src/Pos/Sinbin/Shutdown/Logic.hs
@@ -1,24 +1,0 @@
-module Pos.Sinbin.Shutdown.Logic
-       ( triggerShutdown
-       , waitForShutdown
-       ) where
-
-import           Universum
-
-import           Control.Concurrent.STM (check, readTVar, writeTVar)
-import           System.Wlog (WithLogger, logInfo)
-
-import           Pos.Sinbin.Shutdown.Class (HasShutdownContext (..))
-import           Pos.Sinbin.Shutdown.Types (ShutdownContext (..),
-                     shdnIsTriggered)
-
-triggerShutdown
-    :: (MonadIO m, MonadReader ctx m, WithLogger m, HasShutdownContext ctx)
-    => m ()
-triggerShutdown = do
-    logInfo "NODE SHUTDOWN TRIGGERED, WAITING FOR WORKERS TO TERMINATE"
-    view (shutdownContext . shdnIsTriggered) >>= atomically . flip writeTVar True
-
--- | Wait for the shutdown var to be true.
-waitForShutdown :: ShutdownContext -> IO ()
-waitForShutdown (ShutdownContext v) = atomically (readTVar v >>= check)
